Update python/mlx/nn/layers/normalization.py
Update BatchNorm to support NLC and NHWC input formats

In our convolution operations, we follow the convention that the channels are the last dimension. This commit updates the BatchNorm implementation to support inputs where the channels are the last dimension (NLC or NHWC). This involves changing the dimensions we normalize over and the dimensions we expand our parameters over.

Co-authored-by: Robert McCraith <mccraithrobert@gmail.com>