[Test] Small vector optimization for array inputs
This PR means to be used as an evaluation of the possible performance
gain instead of something to be merged. Read following discussion to get
the background:
https://github.com/ml-explore/mlx/pull/895#issuecomment-2021606855

In summary, using std::vector to store array and it has 2 problems:
1. std::vector does heap allocation for even 1 element.
2. aggregate constructor std::vector{a1, a2, a3} copies args.

I wrote a very trivial small_vector implementation, which stores
elements on stack when aggregate constructor is used, and converts to
std::vector when needed.

In this way, when using the aggregate constructor there is no heap
allocation happened and the args are moved instead of copied.

There is no need to convert the whole codebase to use small_vector, it
is designed to work with existing code with minimal overhead, so we can
replace the std::vector<array> with small_vector<array> piece by piece.

The overhead should be invisible: the small_vector is a thin wrapper of
std::variant<std::array, std::vector>, and no penalty for accessing
elements. When calling eval/vjp/jvp the small_vector is converted to
std::vector by copying values from stack storage to std::vector, which
is expensive but does not make things worse, because the copy would be
done anyway when using the aggregate constructor of std::vector. And we
should be able to remove most of the stack to vector copies eventually.

Currently small_vector is designed for computing graphs, but it should
also be possible to be used for storing shapes.

There are some large changes in array.h, because to use small_vector the
ArrayDesc must know the size of array so it can no longer be a sub-class
in array, and since ArrayDesc becomes opaque to array, the array's
methods can no longer be defined inside the class definition but must be
moved out.