Add use_optimal_threadgroups optional flag to run custom kernel with optimal occupancy
This flag can be used to use dispatch_threadgroups instead of dispatch_threads (the default) to execute the custom kernel. If True, the optimal grid and threadgroup will be calculated for optimal occupancy. However,
the programming of the kernel must assume a flattened operation equivalent to a new grid specification with [grid[0]*grid[1]*grid[2],1,1]. Default: ``False``